requirements.txt
fastapi
python-dotenv
httpx
numpy
uvicorn